Change Notifier to send reference instead of copy
Before, both Notifier sub-classes were sending a copy of the current
snapshot so that no subscriber could modify it. But it was only a
shallow copy and so there were still references, and the Notifier class
would need to know if the data was an regular object or an array. By
sending a reference and being careful not to modify the snapshot
elsewhere, the snapshot doesn't need to be deep-cloned, saving memory.