Fix failure to detect non-existent gid in GamePage
In order to detect the case where a user edits the url parameter
manually to send an invalid gid to GamePage, try to listen to any game
requested of the Firebase listener, so that it can return a null entry
back to GamePage.

If we could guarantee that GamePage would only make a game listener
registration request after the Firebase listener knows the user's games,
this requirement could be avoided. Another way to accomplish this would
be for GamePage to register to listen to user, and then only register
to listen to game once the gid was verified to belong to the user.